Remove useless 'use lib'
Unit tests do not need to add 'use lib "lib"',
running 'prove -l' will automatically add it for you.

Concerning the two other scripts, it's unsafe to add 'lib'
to @INC, as this could result from loading .pm files from
any location. (depending from where the command is run and
which user is running it).

Once installed, it will use the default Perl @INC locations.
For development purpose, adding a -Ilib is probably better.

Using FindBin can be aceptable but also leaves a hole as the
lib directory might belong to a different user in production.